using more custom components
