--- 26---A-Pretty-Deep-Dive-Introduction-to-Next.js/05-onwards-foodies-starting-project/app/components/main-header.js
@@ -1,24 +1,0 @@
-import Link from "next/link";
-import Image from "next/image";
-import logoImg from "@/assets/logo.png";
-import classes from "./main-header.module.css";
-
-export default function MainHeader() {
-    return <header className={classes.header}>
-        <Link className={classes.logo} href="">
-            <Image src={logoImg} alt="A plate with food on it" priority />
-            NextLevel food
-        </Link>
-
-        <nav className={classes.nav}>
-            <ul>
-                <li>
-                    <Link href="/meals">Browse Meals</Link>
-                </li>
-                <li>
-                    <Link href="/community">Foodies Community</Link>
-                </li>
-            </ul>
-        </nav>
-    </header>;
-}
