+++ 26---A-Pretty-Deep-Dive-Introduction-to-Next.js/05-onwards-foodies-starting-project/app/globals.css
@@ -13,20 +13,6 @@ body {
   background: radial-gradient(#282c34, #282c34);
 }
 
-.header-background {
-  position: absolute;
-  width: 100%;
-  height: 320px;
-  top: 0;
-  left: 0;
-  z-index: -1;
-}
-
-svg {
-  display: block;
-  width: 100%;
-  height: auto;
-}
 
 .not-found,
 .error {
